scipy gets its own module.
